Extract `JavaFuzzingContext` into separate interface
